+++ scripts/base-config.php
@@ -70,7 +70,6 @@ return [
 		'SecurityCheck-RCE',
 		'SecurityCheck-PathTraversal',
 		'SecurityCheck-ReDoS',
-		'SecurityCheck-OTHER',
 		// Rely on severity setting to prevent false positive.
 		'SecurityCheck-LikelyFalsePositive',
 		'PhanSyntaxError',