+++ src/SecurityCheckPlugin.php
@@ -94,15 +94,11 @@ abstract class SecurityCheckPlugin extends PluginV3 implements
 	public const REGEX_TAINT = 1 << 14;
 	public const REGEX_EXEC_TAINT = 1 << 15;
 
-	// For stuff that doesn't fit another category
-	public const MISC_TAINT = 1 << 16;
-	public const MISC_EXEC_TAINT = 1 << 17;
-
 	// To allow people to add other application specific taints.
-	public const CUSTOM1_TAINT = 1 << 18;
-	public const CUSTOM1_EXEC_TAINT = 1 << 19;
-	public const CUSTOM2_TAINT = 1 << 20;
-	public const CUSTOM2_EXEC_TAINT = 1 << 21;
+	public const CUSTOM1_TAINT = 1 << 16;
+	public const CUSTOM1_EXEC_TAINT = 1 << 17;
+	public const CUSTOM2_TAINT = 1 << 18;
+	public const CUSTOM2_EXEC_TAINT = 1 << 19;
 
 	// Special purpose for supporting MediaWiki's IDatabase::select
 	// and friends. Like SQL_TAINT, but only applies to the numeric
@@ -112,12 +108,12 @@ abstract class SecurityCheckPlugin extends PluginV3 implements
 	// The associative keys also have this flag if they are tainted.
 	// It is also assumed anything with this flag will also have
 	// the SQL_TAINT flag set.
-	public const SQL_NUMKEY_TAINT = 1 << 22;
-	public const SQL_NUMKEY_EXEC_TAINT = 1 << 23;
+	public const SQL_NUMKEY_TAINT = 1 << 20;
+	public const SQL_NUMKEY_EXEC_TAINT = 1 << 21;
 
 	// For double escaped variables
-	public const ESCAPED_TAINT = 1 << 24;
-	public const ESCAPED_EXEC_TAINT = 1 << 25;
+	public const ESCAPED_TAINT = 1 << 22;
+	public const ESCAPED_EXEC_TAINT = 1 << 23;
 
 	// Special purpose flags (Starting at 2^28)
 	// TODO Renumber these. Requires changing format of the hardcoded arrays
@@ -139,8 +135,7 @@ abstract class SecurityCheckPlugin extends PluginV3 implements
 
 	// YES_TAINT denotes all taint a user controlled variable would have
 	public const YES_TAINT = self::HTML_TAINT | self::SQL_TAINT | self::SHELL_TAINT | self::SERIALIZE_TAINT |
-		self::PATH_TAINT | self::CODE_TAINT | self::REGEX_TAINT | self::CUSTOM1_TAINT | self::CUSTOM2_TAINT |
-		self::MISC_TAINT;
+		self::PATH_TAINT | self::CODE_TAINT | self::REGEX_TAINT | self::CUSTOM1_TAINT | self::CUSTOM2_TAINT;
 	public const EXEC_TAINT = self::YES_TAINT << 1;
 	// @phan-suppress-next-line PhanUnreferencedPublicClassConstant
 	public const YES_EXEC_TAINT = self::YES_TAINT | self::EXEC_TAINT;
@@ -389,7 +384,6 @@ abstract class SecurityCheckPlugin extends PluginV3 implements
 			self::CODE_TAINT => 'CODE',
 			self::PATH_TAINT => 'PATH',
 			self::REGEX_TAINT => 'REGEX',
-			self::MISC_TAINT => 'MISC',
 			self::SQL_NUMKEY_TAINT => 'SQL_NUMKEY',
 			self::ARRAY_OK => 'ARRAY_OK',
 			self::ALL_EXEC_TAINT => '*ALL',
@@ -403,7 +397,6 @@ abstract class SecurityCheckPlugin extends PluginV3 implements
 			self::CODE_EXEC_TAINT => '*CODE',
 			self::PATH_EXEC_TAINT => '*PATH',
 			self::REGEX_EXEC_TAINT => '*REGEX',
-			self::MISC_EXEC_TAINT => '*MISC',
 			self::SQL_NUMKEY_EXEC_TAINT => '*SQL_NUMKEY',
 		];
 
@@ -567,7 +560,7 @@ abstract class SecurityCheckPlugin extends PluginV3 implements
 	 * doesn't seem needed.
 	 *
 	 * The following keywords are supported where {type} can be
-	 * html, sql, shell, serialize, custom1, custom2, misc, sql_numkey,
+	 * html, sql, shell, serialize, custom1, custom2, sql_numkey,
 	 * escaped.
 	 *  * {type} - just set the flag. 99% you should only use 'none' or 'tainted'
 	 *  * exec_{type} - sets the exec flag.
@@ -595,7 +588,7 @@ abstract class SecurityCheckPlugin extends PluginV3 implements
 	 */
 	public static function parseTaintLine( string $line ): ?array {
 		$types = '(?P<type>htmlnoent|html|sql|shell|serialize|custom1|'
-			. 'custom2|misc|code|path|regex|sql_numkey|escaped|none|tainted)';
+			. 'custom2|code|path|regex|sql_numkey|escaped|none|tainted)';
 		$prefixes = '(?P<prefix>escapes|onlysafefor|exec)';
 		$taintExpr = "(?P<taint>(?:${prefixes}_)?$types|array_ok|allow_override)";
 
@@ -678,7 +671,7 @@ abstract class SecurityCheckPlugin extends PluginV3 implements
 	 *
 	 * @note htmlnoent treated like self::HTML_TAINT.
 	 * @param string $name one of:
-	 *   html, sql, shell, serialize, custom1, custom2, code, path, regex, misc, sql_numkey,
+	 *   html, sql, shell, serialize, custom1, custom2, code, path, regex, sql_numkey,
 	 *   escaped, none (= self::NO_TAINT), tainted (= self::YES_TAINT)
 	 * @return int One of the TAINT constants
 	 */
@@ -703,8 +696,6 @@ abstract class SecurityCheckPlugin extends PluginV3 implements
 				return self::PATH_TAINT;
 			case 'regex':
 				return self::REGEX_TAINT;
-			case 'misc':
-				return self::MISC_TAINT;
 			case 'sql_numkey':
 				return self::SQL_NUMKEY_TAINT;
 			case 'escaped':