+++ tests/integration-test-config.php
@@ -241,7 +241,6 @@ return [
 		'SecurityCheck-RCE',
 		'SecurityCheck-PathTraversal',
 		'SecurityCheck-ReDoS',
-		'SecurityCheck-OTHER',
 		'SecurityCheck-LikelyFalsePositive',
 		'SecurityCheckDebugTaintedness',
 		'SecurityCheckInvalidAnnotation',